🪾 mock_data => 🛠️ uploaded thumbnails

+++ apps/server/app/src/main/java/server/decorators/types/AppFile.java
@@ -16,6 +16,7 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import lombok.Getter;
 import server.conf.cloud.etc.data_structure.CloudResourceT;
 import server.decorators.core.ErrAPI;
+import server.lib.data_structure.LibRuntime;
 import server.lib.paths.LibPath;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY, getterVisibility = JsonAutoDetect.Visibility.NONE, isGetterVisibility = JsonAutoDetect.Visibility.NONE)
@@ -42,16 +43,15 @@ public final class AppFile {
     }
 
     // ? this constructor below is used mostly for dev uploads of mock data images
-    // public AppFile(
-    //         String filename,
-    //         String contentType,
-    //         Path filePath) {
-    //     this.field = "images";
-    //     this.filename = randomFilename(filename);
-    //     this.contentType = "";
-    //     this.bts = LibRuntime.inTryBlock(() -> Files.readAllBytes(filePath));
-    //     this.filePath = filePath;
-    // }
+    public AppFile(
+            String filename,
+            Path filePath) {
+        this.field = CloudResourceT.IMAGE;
+        this.filename = randomFilename(filename);
+        this.contentType = Nullable.asNone();
+        this.bts = Nullable.of(LibRuntime.inTryBlock(() -> Files.readAllBytes(filePath)));
+        this.filePath = Nullable.of(filePath);
+    }
 
     private final String findExt(String arg) {
         final String ext;
@@ -65,7 +65,7 @@ public final class AppFile {
     }
 
     private final String randomFilename(String arg) {
-        String ext = findExt(arg);
+        final String ext = findExt(arg);
         return UUID.randomUUID().toString() + ext;
     }
 
@@ -74,6 +74,9 @@ public final class AppFile {
             throw new ErrAPI("tried to save locally without filepath present");
         if (bts.isNone())
             throw new ErrAPI("tried to save locally without binary code present");
+
+        if (!Files.exists(filePath.orYell()))
+            throw new ErrAPI("file does not exist at => " + filePath.orYell());
 
         try {
             Files.write(this.getFilePath().orYell(), this.getBts().orYell(),
@@ -85,7 +88,6 @@ public final class AppFile {
     }
 
     public final void deleteLocally() {
-
         if (filePath.isNone())
             throw new ErrAPI("tried to delete locally without filepath present");
 
@@ -110,7 +112,7 @@ public final class AppFile {
 
     public final FileSystemResource getResourceFromPath() {
         if (filePath.isNone())
-            throw new ErrAPI("tried to create resource from file path without filepath present");
+            throw new ErrAPI("tried to create resource from filepath without filepath present");
 
         return new FileSystemResource(filePath.orYell());
     }
@@ -127,7 +129,7 @@ public final class AppFile {
                 final Object val = f.get(this);
 
                 if ("bts".equals(f.getName()))
-                    fancyMap.put("bytes", "💾 long binary code...");
+                    fancyMap.put("bytes", val);
                 else
                     fancyMap.put(f.getName(), val instanceof Nullable<?> inst ? inst.orNone() : val);
